+++ src/main/java/com/MyMongoSpring/MyMongoSpring/Repositary/StudentRepo.java
@@ -3,5 +3,5 @@ package com.MyMongoSpring.MyMongoSpring.Repositary;
 import com.MyMongoSpring.MyMongoSpring.Model.Student;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface StudentRepo extends MongoRepository<Student,Integer> {
+public interface StudentRepo extends MongoRepository<Student,Long> {
 }
